backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==5.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
APScheduler==3.10.4
groq==0.4.2
requests==2.31.0
python-binance==1.0.17
pandas==2.1.1
numpy==1.24.3
gunicorn==21.2.0
